done review admin + jseeker + update role for deleting something

+++ employer_JM_delete.php
@@ -5,13 +5,19 @@
 	$model = new Model;
 
 	$jobid = $_GET['id'];
-	$delete = $model->deleteJob($jobid);
+    if (isset($_SESSION['role']) && $_SESSION['role'] == 3) {
+	    $delete = $model->deleteJob($jobid);
+    }else{
+        $delete = false;
+    }
 
     if($delete){
 
         echo "<script>alert('delete successfully');</script>";
         echo "<script>window.location.href = 'employer_JM.php';</script>";
 
+    }else{
+        echo "<script>alert('You're not suppose to do this');</script>";
     }
 
 ?>
